build: enable c++23 for core
Due to cocos2d-x uses syntax that is not accepted by c++ 20 any more,
this change compiles core separately, and links it to the final
exectuable.

Note that core does not links against cocos2d-x so after this change we
will lose access to cocos2d-x stuff when writing core.